Divided into files.
